refactor(plugin): align shared modules with plugin boundaries
- Move RouterOS support from infra into the executor namespace.
- Rename matcher and provider utility modules for clearer ownership.
- Preserve runtime behavior while updating module declarations and imports.
- Verify the refactor with the full quality checks.
