+++ src/plugin/executor/mod.rs
@@ -74,6 +74,8 @@ pub mod reverse_lookup;
 pub mod ros_address_list;
 #[cfg(feature = "plugin-ros-route")]
 pub mod ros_route;
+#[cfg(any(feature = "plugin-ros-address-list", feature = "plugin-ros-route"))]
+pub(crate) mod routeros;
 #[cfg(feature = "plugin-script")]
 pub mod script;
 pub mod sequence;
