+++ src/plugin/executor/ros_route/api.rs
@@ -15,8 +15,8 @@ use tracing::warn;
 
 use super::model::{RouteCommentCodec, RouteFamily, RouteKey};
 use crate::infra::error::{DnsError, Result};
-use crate::infra::mikrotik::batching::join_all_bounded;
-use crate::infra::mikrotik::transport::{
+use crate::plugin::executor::routeros::batching::join_all_bounded;
+use crate::plugin::executor::routeros::transport::{
     RouterOsConnectionConfig, RouterOsEvent, RouterOsResult, RouterOsTimeouts, RouterOsTransport,
     RouterOsTransportSnapshot,
 };
