+++ src/plugin/executor/routeros/completion.rs
