+++ src/plugin/matcher/mod.rs
@@ -30,7 +30,6 @@ pub mod false_matcher;
 pub mod has_resp;
 pub mod has_wanted_ans;
 pub mod mark;
-pub mod matcher_utils;
 pub mod ptr_ip;
 pub mod qclass;
 pub mod qname;
@@ -40,6 +39,7 @@ pub mod random;
 pub mod rate_limiter;
 pub mod rcode;
 pub mod resp_ip;
+pub mod rules;
 pub mod string_exp;
 pub mod time;
 pub mod true_matcher;
